Infer attribute type from value (#62)
* Add descriptive error messages

* Exclude examples directory from language statistics
Prevents the repo from showing up as written in Java instead of Rust
just because it contains an Android App demo.

* Inter attribute type based on value

Previously, these were hardcoded but it seems the parser might just take
a guess, and we can too. This greatly expands support for properties in
the Manifest we haven't seen before.