Fix new clippy lint (#65)
* Add descriptive error messages

* Exclude examples directory from language statistics
Prevents the repo from showing up as written in Java instead of Rust
just because it contains an Android App demo.

* Inter attribute type based on value

Previously, these were hardcoded but it seems the parser might just take
a guess, and we can too. This greatly expands support for properties in
the Manifest we haven't seen before.

* Fix new clippy lint

There was a new lint added to clippy that recommends putting logged
objects directly in format strings, and this was causing CI to fail for
all changes.

Also fix tiny typo.

* Reformat after lint fix

Cargo fix doesn't format afterward